:sparkles: feat: Validações referentes à autenticação na entidade especialidade

+++ src/main/java/com/connectCare/connectCareApi/services/impl/EspecialidadeServiceImpl.java
@@ -1,14 +1,18 @@
 package com.connectCare.connectCareApi.services.impl;
 
 import com.connectCare.connectCareApi.exceptions.EspecialidadeNaoEncontradaException;
+import com.connectCare.connectCareApi.exceptions.NaoAutorizadoException;
 import com.connectCare.connectCareApi.exceptions.NenhumRegistroEncontradoException;
 import com.connectCare.connectCareApi.exceptions.OperacaoBancoDeDadosException;
 import com.connectCare.connectCareApi.models.entities.Especialidade;
 import com.connectCare.connectCareApi.repositories.EspecialidadeRepository;
 import com.connectCare.connectCareApi.services.GenericService;
+import com.connectCare.connectCareApi.utils.UsuarioAutenticado;
+
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,6 +25,11 @@ public class EspecialidadeServiceImpl implements GenericService<Especialidade> {
 
     @Override
     public Especialidade create(Especialidade especialidade) {
+    	//Verificando se o atual usuário é ADMIN
+        if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+            throw new NaoAutorizadoException();
+        }
+        
         return repository.save(especialidade);
     }
 
@@ -41,6 +50,11 @@ public class EspecialidadeServiceImpl implements GenericService<Especialidade> {
     public Especialidade update(Especialidade especialidade) {
         try{
             Especialidade especialidadeEncontrada = repository.getReferenceById(especialidade.getId());
+            
+            //Verificando se o atual usuário é ADMIN
+            if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+                throw new NaoAutorizadoException();
+            }
 
             especialidadeEncontrada.setNome(especialidade.getNome());
 
@@ -56,6 +70,12 @@ public class EspecialidadeServiceImpl implements GenericService<Especialidade> {
     public void delete(Integer id) {
         try{
             Especialidade especialidadeEncontrada = repository.findById(id).orElseThrow(() -> new EspecialidadeNaoEncontradaException(id));
+            
+            //Verificando se o atual usuário é ADMIN
+            if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+                throw new NaoAutorizadoException();
+            }
+            
             repository.delete(especialidadeEncontrada);
         }catch(DataIntegrityViolationException e) {
             throw new OperacaoBancoDeDadosException("Não foi possível excluir, pois essa especialidade está relacionada com algum médico.");
